Implementação do Menu Completo
