+++ src/App.js
@@ -1,5 +1,5 @@
 import './App.css';
-import Topo from './components/topo';
+import Topo from './components/Topo';
 
 function App() {
   return (
