+++ src/components/Topo.js
@@ -1,0 +1,17 @@
+import NavMenu from './NavMenu';
+import Options from './Options'
+import  'bootstrap/dist/css/bootstrap.min.css' ;
+import './style/topo.css'
+
+function Topo() {
+    return (
+      <>
+      <NavMenu />
+      <Options />
+      </>
+    );
+  }
+  
+  export default Topo;
+
+
