+++ src/components/style/navmenu.css
@@ -1,9 +1,3 @@
-@import url('https://fonts.googleapis.com/css2?family=Roboto:wght@300;400;700&display=swap');
-
-
-*{
-  font-family: 'Roboto', sans-serif;
-}
 .NavBar-Logo{
     padding-left: 5px;
 }
