+++ src/components/style/options.css
@@ -1,0 +1,3 @@
+.navbar-dark .navbar-nav .nav-link{
+    color: #222222;
+}
